refactor: 수정/삭제 alert 로직 수정

+++ src/main/resources/static/js/post/post_read.js
@@ -1,5 +1,6 @@
 let process = 'none';
 const idx = $('#postIdx').val();
+const password = $('#password');
 const Read = {
   init: function () {
     let innerThis = this;
@@ -9,29 +10,29 @@ const Read = {
     });
     $('.btn_modify').on('click', function () {
       $('.verify_box').css('visibility', 'visible');
-      $('#password').focus();
       process = 'modify';
+      innerThis.emptyBox();
     });
     $('.btn_delete').on('click', function () {
       $('.verify_box').css('visibility', 'visible');
       process = 'delete';
+      innerThis.emptyBox();
     })
     $('.verify_cancel').on('click', function () {
       $('.verify_box').css('visibility', 'hidden');
     })
     $('#password').on('keyup', function (event) {
       if (event.keyCode === 13) {
-        if (confirm(`${innerThis.comment()}`)) {
-          isPWCorrect = innerThis.verify();
-          if (isPWCorrect && process === 'modify') {
-            innerThis.modify();
-          }
-          if (isPWCorrect && process === 'delete') {
-            innerThis.delete();
-          }
-          if (!isPWCorrect) {
-            alert('비밀번호가 일치하지 않습니다. 다시 확인해주세요.');
-          }
+        isPWCorrect = innerThis.verify();
+        if (isPWCorrect && process === 'modify') {
+          innerThis.modify();
+        }
+        if (isPWCorrect && process === 'delete') {
+          innerThis.delete();
+        }
+        if (!isPWCorrect) {
+          alert('비밀번호가 일치하지 않습니다. 다시 확인해주세요.');
+          password.select();
         }
       }
     })
@@ -41,7 +42,7 @@ const Read = {
     let isPWCorrect
     let data = {
       postIdx: idx,
-      password: $('#password').val()
+      password: password.val()
     };
     $.ajax({
       type: 'POST',
@@ -66,20 +67,27 @@ const Read = {
     }
   },
 
+  emptyBox: function() {
+    password.val('');
+    password.focus();
+  },
+
   modify: function() {
     location.href = `/post/modify/${idx}`;
   },
 
   delete: function() {
-    $.ajax({
-      type: 'DELETE',
-      url: `/post/delete/${idx}`,
-    }).done(function () {
-      alert(`글이 삭제되었습니다.`);
-      location.href = '/post/list';
-    }).fail(function (error) {
-      alert(`관리자에게 문의하세요. error : ${JSON.stringify(error)}`);
-    })
+    if(confirm(`${this.comment()}`)) {
+      $.ajax({
+        type: 'DELETE',
+        url: `/post/delete/${idx}`,
+      }).done(function () {
+        alert(`글이 삭제되었습니다.`);
+        location.href = '/post/list';
+      }).fail(function (error) {
+        alert(`관리자에게 문의하세요. error : ${JSON.stringify(error)}`);
+      });
+    }
   }
 }
